[refactor] Centralized server auth code
