+++ mme_server/datastore.py
@@ -81,123 +81,6 @@ class ESIndex:
             self.refresh()
 
 
-class ServerManager:
-    FIELDS = ['server_id', 'server_label', 'server_key', 'direction', 'base_url']
-    DOC_CONFIG = {
-        'properties': {
-            'server_id': {
-                'type': 'string',
-                'index': 'not_analyzed',
-            },
-            'server_label': {
-                'type': 'string',
-                'index': 'not_analyzed',
-            },
-            'server_key': {
-                'type': 'string',
-                'index': 'not_analyzed',
-            },
-            'direction': {
-                'type': 'string',
-                'index': 'not_analyzed',
-            },
-            'base_url': {
-                'type': 'string',
-                'index': 'not_analyzed',
-            }
-        }
-    }
-
-    def __init__(self, backend=None):
-        if backend is None:
-            backend = Elasticsearch()
-
-        self._index = ESIndex(db=backend,
-                              name='servers',
-                              doc_type='server',
-                              doc_config=self.DOC_CONFIG)
-
-    @property
-    def index(self):
-        return self._index
-
-    def add(self, server_id, server_label, server_key, direction, base_url):
-        assert server_id and server_key and direction in ['in', 'out']
-
-        if base_url and not base_url.startswith('https://'):
-            logger.error('base URL must start with "https://"')
-            return
-
-        self.index.ensure_exists()
-
-        data = {
-            'server_id': server_id,
-            'server_label': server_label,
-            'server_key': server_key,
-            'direction': direction,
-            'base_url': base_url,
-        }
-
-        # If it already exists, update
-        s = self.index.search()
-        s = s.filter('term', server_id=server_id)
-        s = s.filter('term', direction=direction)
-        results = s.execute()
-
-        if results.hits.total > 1:
-            logger.error('Found two or more matching server entries')
-        else:
-            id = None
-            if results.hits.total == 1:
-                # Found a match, so update instead
-                id = results.hits[0].meta.id
-
-            self.index.save(id=id, doc=data)
-            logger.info("Authorized server:\n{}".format(json.dumps(data, indent=4, sort_keys=True)))
-            # Refresh index to ensure immediately usable
-            self.index.refresh()
-
-    def remove(self, server_id, direction):
-        if self.index.exists():
-            s = self.index.search()
-            s = s.filter('term', server_id=server_id)
-
-            if direction:
-                s = s.filter('term', direction=direction)
-            results = s.execute()
-
-            for hit in results:
-                id = hit.meta.id
-                self.index.delete(id=id)
-                logger.info("Deleted server:{} direction:{}".format(hit.server_id, hit.direction))
-
-    def list(self):
-        rows = []
-        if self.index.exists():
-            s = self.index.search()
-            s = s.query('match_all')
-
-            # Iterate through all, using scan
-            for hit in s.scan():
-                row = dict([(field, hit[field]) for field in self.FIELDS])
-                rows.append(row)
-
-        return {
-            'fields': self.FIELDS,
-            'rows': rows
-        }
-
-    def verify(self, key):
-        if key and self.index.exists():
-            s = self.index.search()
-            s = s.filter('term', server_key=key)
-            s = s.filter('term', direction='in')
-            results = s.execute()
-
-            if results.hits:
-                return results.hits[0]
-
-
 class PatientManager:
     DOC_CONFIG = {
         '_all': {
@@ -392,4 +275,3 @@ class DatastoreConnection:
         self._db = Elasticsearch()
         self.patients = PatientManager(backend=self._db)
         self.vocabularies = VocabularyManager(backend=self._db)
-        self.servers = ServerManager(backend=self._db)
